backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg[binary]
psycopg_pool
